BE - change architecture to properly add components and templates
